+++ cmd/project/list.go
@@ -7,6 +7,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/prolific-oss/cli/client"
+	"github.com/prolific-oss/cli/ui"
 	"github.com/spf13/cobra"
 )
 
@@ -14,10 +15,12 @@ import (
 type ListOptions struct {
 	Args        []string
 	WorkspaceID string
+	Limit       int
+	Offset      int
 }
 
 // NewListCommand creates a new command to deal with projects
-func NewListCommand(commandName string, client client.API, w io.Writer) *cobra.Command {
+func NewListCommand(commandName string, c client.API, w io.Writer) *cobra.Command {
 	var opts ListOptions
 
 	cmd := &cobra.Command{
@@ -26,7 +29,7 @@ func NewListCommand(commandName string, client client.API, w io.Writer) *cobra.C
 		RunE: func(cmd *cobra.Command, args []string) error {
 			opts.Args = args
 
-			err := renderProjects(client, opts, w)
+			err := renderProjects(c, opts, w)
 			if err != nil {
 				return fmt.Errorf("error: %s", err.Error())
 			}
@@ -37,6 +40,8 @@ func NewListCommand(commandName string, client client.API, w io.Writer) *cobra.C
 
 	flags := cmd.Flags()
 	flags.StringVarP(&opts.WorkspaceID, "workspace", "w", "", "Filter projects by workspace.")
+	flags.IntVarP(&opts.Limit, "limit", "l", client.DefaultRecordLimit, "Limit the number of workspaces returned")
+	flags.IntVarP(&opts.Offset, "offset", "o", client.DefaultRecordOffset, "The number of workspaces to offset")
 
 	return cmd
 }
@@ -47,7 +52,7 @@ func renderProjects(client client.API, opts ListOptions, w io.Writer) error {
 		return errors.New("please provide a workspace ID")
 	}
 
-	projects, err := client.GetProjects(opts.WorkspaceID)
+	projects, err := client.GetProjects(opts.WorkspaceID, opts.Limit, opts.Offset)
 	if err != nil {
 		return err
 	}
@@ -58,5 +63,9 @@ func renderProjects(client client.API, opts ListOptions, w io.Writer) error {
 		fmt.Fprintf(tw, "%s\t%s\t%v\n", project.ID, project.Title, project.Description)
 	}
 
-	return tw.Flush()
+	_ = tw.Flush()
+
+	fmt.Fprintf(w, "\n%s\n", ui.RenderRecordCounter(len(projects.Results), projects.Meta.Count))
+
+	return nil
 }